requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
docker==6.0.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.89.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
motor==3.1.1
packaging==22.0
pluggy==1.0.0
pydantic==1.10.4
pymongo==4.3.3
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websocket-client==1.4.2
websockets==10.4
